% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linemodels.R
\name{line.models.with.features}
\alias{line.models.with.features}
\title{Feature dependent priors on linemodels}
\usage{
line.models.with.features(
  X,
  SE,
  scales,
  slopes,
  cors,
  model.names = NULL,
  r.lkhood = rep(0, ncol(X) * (ncol(X) - 1)/2),
  n.iter = 200,
  n.burnin = 20,
  features = NULL,
  verbose = TRUE
)
}
\arguments{
\item{X}{nxM matrix of estimates with M columns (one col per effect variable and one row per observation)}

\item{SE}{nxM matrix of standard errors with M columns (one col per effect variable and one row per observation)}

\item{scales}{K-vector of standard deviations of the largest effect, one per each linemodel}

\item{slopes}{Kx(M-1)-matrix of slopes, one row per linemodel
If M = 2, can be given also as a K-vector of slopes, one per each linemodel}

\item{cors}{K-vector of correlation parameters, one per each linemodel}

\item{model.names}{K-vector of names of linemodels}

\item{r.lkhood}{MxM correlation matrix of the estimators of the M effects
can be given also as a vector of the upper triangular values in the row-major order}

\item{n.iter}{number of Gibbs sampler iterations (after burn-in)}

\item{n.burnin}{number of burn-in iterations that will be discarded}

\item{features}{nxp matrix of features with one row per observation and one column per feature.
If NULL, then only a constant feature is modelled.
Every feature much show variation among observations
while an intercept is added to the model automatically.}

\item{verbose}{if TRUE, prints the index of every 100th iteration}
}
\value{
List with two components: (1) 'params', K*(p+1)x4 matrix of
posterior distribution of regression coefficient of each features for each linemodel,
where p is the number of (columns of) 'features'.
Columns are mean, lower and upper points of 95\% credible interval
and standard deviation.
One row per coefficient, each linemodel has p+1 coefficients.
(2)  'groups', nxK matrix with one column per linemodel and one row per observation giving
membership probabilities of each observation in each model.
}
\description{
Consider M effect sizes.
The linemodels are defined by three parameters:
scale = standard deviation of the (largest) effect,
slope = set of M-1 slopes defining the line around which the effects are scattered,
cor = non-negative pairwise correlation between each effect pair,
where cor = 1 means the same effect and cor = 0 means independent effects.
}
\details{
A linemodel has the following properties:

1) Effects are scattered around line defined by vector (1,slope_1,...,slope_(M-1)).
   Each slope_i is the slope between effect variables i+1 and 1.
   Slope can be any real number or +/-Inf.
   If any slope is infinite, the effect of variable 1 and variables with finite slopes are 0.
   It is recommended to keep the slopes finite to avoid problems with interpretation.

2) The largest of the prior variances of the effects is scale^2.
   The effect that has the largest scale is determined by the slopes,
   and the scales of the remaining effects are determined by the slopes as well.

3) Distribution around the line is determined as follows.
   Consider a distribution scattered around the direction of the vector (1,1,...,1)
   with the given constant value of the pairwise correlations between the effect variables.
   Rotate that distribution by an orthogonal rotation that rotates the direction of the
   vector (1,1,...,1) to the direction of the vector (1,slope_1,...,slope_(M-1))
   and use the corresponding distribution, scaled so that the maximum variance
   among the effects is set to scale^2.
   NOTE: This is not same as having the given correlation around line defined by the slopes,
         because the shape of that distribution depends on the slopes but the
        definition we use here is independent of slopes up to an orthogonal transformation.

Examples:
Set scale = 0 to get the null model. (Values of slope and cor do not matter in this case.)
Use scale > 0, slope = 1 and cor = 0 to get independent effects model for 2-dim case.
Use scale > 0, slope = c(1,1,1) and cor = 1 to get fixed effects model for 4-dim case.
Use scale > 0, slope = 1 and 0 < cor < 1  to get correlated effects model for 2-dim case.
              note that cor value needs to be quite near 1 to get very similar values
              between the two effects with high probability (e.g. cor = 0.99...0.999).


The prior distribution of each observation belonging to each linemodel can
depend on observation specific values of 'features' given as parameter.
The relationship between features and prior probabilities is defined by
a logistic/multinomial regression model.
A Gibbs sampler is used for estimating the posterior probabilities of the memberships.
}
\examples{
#2D example
line.models.with.features(
           X = linemodels.ex2[,c("beta1","beta2")],
           SE = linemodels.ex2[,c("se1","se2")],
           scales = c(0.1, 0.1),
           slopes = c(0, 3),
           cors = c(0.999, 0.999),
           features = linemodels.ex2$anno)
#3D example
line.models.with.features(
           X = linemodels.ex2[,c("beta1","beta2","beta3")],
           SE = linemodels.ex2[,c("se1","se2","se3")],
           scales = c(0.2, 0.2),
           slopes = matrix(c(0,-2, 3,6), byrow = TRUE, ncol = 2),
           cors = c(0.999, 0.999),
           features = linemodels.ex2$anno)
}
