% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{linemodels.ex1}
\alias{linemodels.ex1}
\title{linemodels example data set 1}
\format{
\describe{
  \item{beta1, beta2, beta3}{effects on outcomes 1, 2, 3}
  \item{se1, se2, se3}{standard errors of beta estimates}
  \item{maf}{minor allele frequency of variant}
}
}
\usage{
linemodels.ex1
}
\description{
100 genetic variants with associations to three outcomes.
Variants 1..20 from Model1, 21..60 from Model2 and 61..100 from Model3.
All models have scale = 0.2 and cor = 0.995.
Slopes are: Model1 = (0,0), Model2 = (0.5,0.2), Model3 = (1,1).
}
\keyword{datasets}
