% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{linemodels.ex2}
\alias{linemodels.ex2}
\title{linemodels example data set 2}
\format{
\describe{
  \item{beta1, beta2, beta3}{effects on outcomes 1, 2, 3}
  \item{se1, se2, se3}{standard errors of beta estimates}
  \item{annotation}{binary annotation that can be used as a feature}
}
}
\usage{
linemodels.ex2
}
\description{
101 genetic variants with associations to three outcomes.
Variants 1..50 from Model1, 51..100 from Model2 and 101 is an outlier.
All models have scale = 0.1 and cor = 0.999.
Slopes are: Model1 = (0, 2), Model2 = (3, 6).
}
\keyword{datasets}
