% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linemodels.R
\name{slope.for.pair}
\alias{slope.for.pair}
\title{Returns slopes between two variables ii and jj,
that is, the slopes 'b' of the line x_ii = b*x_jj}
\usage{
slope.for.pair(ii, jj, slopes)
}
\arguments{
\item{ii}{index of "y-axis" variable}

\item{jj}{index of "x-axis" variable}

\item{slopes}{matrix of size Kx(M-1), where column 'i' gives the
slopes between variable 'i+1' and variable 1 or a vector of length M-1
M is the number of dimensions
K is the number of linemodels}
}
\value{
slopes for the pair specified for all K models
}
\description{
Returns slopes between two variables ii and jj,
that is, the slopes 'b' of the line x_ii = b*x_jj
}
\examples{
slope.for.pair(3,2, c(2,3))
slope.for.pair(3,2, matrix(c(2,3, 0.5,-1), byrow = TRUE, ncol = 2))
}
