% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linemodels.R
\name{visualize.classification.regions}
\alias{visualize.classification.regions}
\title{Visualize classification regions of 2-dimensional line models}
\usage{
visualize.classification.regions(
  scales,
  slopes,
  cors,
  SE,
  r.lkhood = 0,
  model.priors = rep(1, length(scales)),
  col.palette = NULL,
  add.legend = TRUE,
  xlim = NULL,
  ylim = NULL,
  breakpoints = 200,
  xlab = "VARIABLE1",
  ylab = "VARIABLE2",
  cex.lab = 1,
  cex.axis = 1,
  line.lty = 1,
  region.lty = 2,
  line.lwd = 1,
  region.lwd = 1,
  emphasize.axes = TRUE
)
}
\arguments{
\item{scales}{vector of standard deviations of larger effect of linemodels}

\item{slopes}{vector of slopes of linemodels}

\item{cors}{vector of correlation parameters of linemodels}

\item{SE}{vector of two standard errors for variable1 and variable2, respectively}

\item{r.lkhood}{correlation between estimators of the two values}

\item{model.priors}{vector of (unnormalized) prior probabilities of the two linemodels}

\item{col.palette}{vector of seven colors (with reasonable defaults)}

\item{add.legend}{if TRUE, adds legend on right-hand side of the plot}

\item{xlim, ylim, xlab, ylab, cex.lab, cex.axis}{standard plotting parameters}

\item{breakpoints}{how many breakpoints for defining the grid on each axis}

\item{line.lty}{plotting type for lines (1 = solid, 2 = dashed ...)}

\item{region.lty}{plotting type for 95\% regions (1 = solid, 2 = dashed ...)}

\item{line.lwd}{width of lines}

\item{region.lwd}{width of line for 95\% regions}

\item{emphasize.axes}{if TRUE, coordinate axes are marked with a black line}
}
\value{
none
}
\description{
Colors the plane according to the classification probabilities
of two competing line models.
See help of 'line.models( )' for details about specifying the models.
}
\examples{
visualize.classification.regions(c(0.1,0.2), c(1,0), c(0.995,0.995), c(0.05,0.05))
}
