% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linemodels.R
\name{visualize.scales}
\alias{visualize.scales}
\title{Visualize effect distributions of line models}
\usage{
visualize.scales(
  scales,
  scale.weights = c(1),
  model.names = NULL,
  model.cols = NULL,
  legend.position = "topleft"
)
}
\arguments{
\item{scales}{vector of standard deviations of larger effect of each linemodel}

\item{scale.weights}{vector of weights of each component or a matrix
with one row per linemodel and one column per component}

\item{model.names}{vector of names of each linemodel}

\item{model.cols}{vector of colors for each linemodel}

\item{legend.position}{position of legend, value NULL removes legend}
}
\value{
none
}
\description{
Plots density functions of distributions of effect sizes
of line models specified by scales and scale.weights.
See help of 'line.models( )' for details about specifying the models.
}
\examples{
visualize.scales(c(0.1,0.2), c(1,0,0,1))
visualize.scales(c(0.1,0.2),
                 rbind(c(1,0,0,1),
                       c(3,0,0,1)))
}
