% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linemodels.R
\name{line.models}
\alias{line.models}
\title{Membership probabilities in line models}
\usage{
line.models(
  X,
  SE,
  scales,
  slopes,
  cors,
  model.names = NULL,
  model.priors = rep(1/length(slopes), length(slopes)),
  r.lkhood = 0,
  scale.weights = c(1)
)
}
\arguments{
\item{X}{matrix of effect sizes with two columns (one col per trait)}

\item{SE}{matrix of standard errors with two columns (one col per trait)}

\item{scales}{vector of standard deviations of larger effect of each model}

\item{slopes}{vector of slopes of each model}

\item{cors}{vector of correlation parameters of each model}

\item{model.names}{vector of names of each model}

\item{model.priors}{vector of prior probabilities of models up to a normalization constant}

\item{r.lkhood}{correlation between estimators of the two effects}

\item{scale.weights}{vector of weights of each component or a matrix
with one row per model and one column per component}
}
\value{
matrix with one column per model and one row per variable giving
membership probabilities of each variable in each model
}
\description{
Consider two effects:  x = effect1 and y = effect2.
The line models are defined by three parameters:
scale = standard deviation of the (larger) effect,
slope = slope of line around which the effects are scattered,
cor = non-negative correlation between effects where cor = 1 means equal effect size.
}
\details{
A line model has the following properties:

1) effects are scattered around line y = slope*x.
   'slope' can be any real number or Inf (in which case effect x is zero).

2) the larger of the prior variances of effects is scale^2.
   That is, if |slope| <= 1, then Var(x) = scale^2 and
            if |slope| > 1, then Var(y) = scale^2.

3) Distribution around the line is determined as follows.
   Consider a distribution scattered around line y = x with the given correlation.
   Rotate that distribution by an orthogonal rotation defined by angle
   theta = atan(slope) - pi/4
   and use the corresponding distribution, scaled so that the maximum variance
   of the two effects is scale^2.
   NOTE: This is not same as having the given correlation around line y = slope*x,
         because the shape of that distribution depends on the slope but
         definition we use here is independent of slope (up to an orthogonal transformation).

Effect distribution for the larger effect can be specified as a mixture
of Gaussians.
If 'scale.weights' is a vector of length C > 1, then model 'M' is a mixture
of C two-dimensional Gaussians where component i has mean of 0,
the standard deviation for the larger effect of 'scales[M]/(C - i + 1)'
and the mixture weight is proportional to scale.weights[M].
If 'scale.weights' is a matrix with C > 1 columns and K rows,
then model 'M' is a mixture of 'C' Gaussians specified by row 'M'
of 'scale.weights'.
}
\examples{
line.models(X = linemodels.ex1[,1:2], SE = linemodels.ex1[,3:4],
           scales = c(0.2, 0.2, 0.2),
           slopes = c(0, 0.5, 1),
           cors = c(0.995, 0.995, 0.995))
}
