% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linemodels.R
\name{line.models.with.proportions}
\alias{line.models.with.proportions}
\title{Proportion parameters and membership probabilities of line models}
\usage{
line.models.with.proportions(
  X,
  SE,
  scales,
  slopes,
  cors,
  model.names = NULL,
  r.lkhood = 0,
  n.iter = 200,
  n.burnin = 20,
  diri.prior = rep(1/length(scales), length(scales)),
  verbose = TRUE
)
}
\arguments{
\item{X}{matrix of effect sizes with two columns (one col per trait)}

\item{SE}{matrix of standard errors with two columns (one col per trait)}

\item{scales}{vector of standard deviations of larger effect of each model}

\item{slopes}{vector of slopes of each model}

\item{cors}{vector of correlation parameters of each model}

\item{model.names}{vector of names of each model}

\item{r.lkhood}{correlation between estimators of the two effects}

\item{n.iter}{number of Gibbs sampler iterations (after burn-in)}

\item{n.burnin}{number of burn-in iterations that will be discarded}

\item{diri.prior}{parameters of the Dirichlet distribution used as prior for proportions}

\item{verbose}{if TRUE, prints the index of every 100th iteration}
}
\value{
List with two components: (1) 'params', matrix of
posterior distribution of proportion parameters for each model.
Columns are mean, lower and upper points of 95\% credible interval
and standard deviation. (2)  'groups',
matrix with one column per model and one row per variable giving
membership probabilities of each variable in each model.
}
\description{
Consider two effects:  x = effect1 and y = effect2.
The line models are defined by three parameters:
scale = standard deviation of the (larger) effect,
slope = slope of line around which the effects are scattered,
cor = non-negative correlation between effects,
where cor = 1 means the same effect and cor = 0 means independent effects.
}
\details{
A line model has the following properties:

1) effects are scattered around line y = slope*x.
   'slope' can be any real number or Inf (in which case effect x is zero).

2) the larger of the prior variances of effects is scale^2.
   That is, if |slope| <= 1, then Var(x) = scale^2 and
            if |slope| > 1, then Var(y) = scale^2.

3) Distribution around the line is determined as follows.
   Consider a distribution scattered around line y = x with the given correlation.
   Rotate that distribution by an orthogonal rotation defined by angle
   theta = atan(slope) - pi/4
   and use the corresponding distribution, scaled so that the maximum variance
   of the two effects is scale^2.
   NOTE: This is not same as having the given correlation around line y = slope*x,
         because the shape of that distribution depends on the slope but
         definition we use here is independent of slope (up to an orthogonal transformation).

The prior distribution of the mixture proportions is Dirichlet(diri.prior) and
a Gibbs sampler is used for estimating the posterior.
}
\examples{
line.models.with.proportions(
           X = linemodels.ex1[,1:2], SE = linemodels.ex1[,3:4],
           scales = c(0.2, 0.2, 0.2),
           slopes = c(0, 0.5, 1),
           cors = c(0.995, 0.995, 0.995))
}
