% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{linemodels.ex1}
\alias{linemodels.ex1}
\title{linemodels example data set 1}
\format{
\describe{
  \item{beta1, beta2}{effects on outcomes 1 and 2}
  \item{se1, se2}{standard errors of beta1 and beta2}
}
}
\usage{
linemodels.ex1
}
\description{
100 variables, of which 1..20 originating from model (scale = 0.2, slope = 0, cor = 0.995)
21..60 from (0.2, 0.5, 0.995) and 61..100 from (0.2, 1, 0.995).
}
\keyword{datasets}
