% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linemodels.R
\name{visualize.line.models}
\alias{visualize.line.models}
\title{Visualize linemodels}
\usage{
visualize.line.models(
  scales,
  slopes,
  cors,
  model.names = NULL,
  model.cols = NULL,
  legend.position = "bottomright",
  xlim = NULL,
  ylim = NULL,
  xlab = "VARIABLE1",
  ylab = "VARIABLE2",
  cex.lab = 1,
  cex.axis = 1,
  line.lty = 1,
  region.lty = 2,
  line.lwd = 1,
  region.lwd = 1,
  plot.grid = TRUE,
  plot.new = TRUE,
  emphasize.axes = TRUE
)
}
\arguments{
\item{scales}{vector of standard deviations of larger effect of each linemodel}

\item{slopes}{vector of slopes of each linemodel}

\item{cors}{vector of correlation parameters of each linemodel}

\item{model.names}{vector of names of each linemodel}

\item{model.cols}{vector of colors for each linemodel}

\item{legend.position}{position of legend, value NULL removes legend}

\item{xlim, ylim, xlab, ylab, cex.lab, cex.axis}{standard plotting parameters}

\item{line.lty}{plotting type for lines (1 = solid, 2 = dashed ...)}

\item{region.lty}{plotting type for 95\% regions (1 = solid, 2 = dashed ...)}

\item{line.lwd}{width of lines}

\item{region.lwd}{width of line for 95\% regions}

\item{plot.grid}{if TRUE, plots a grid}

\item{plot.new}{if TRUE, makes a new plot, if FALSE adds to existing plot}

\item{emphasize.axes}{if TRUE, coordinate axes are marked with a black line}
}
\value{
none
}
\description{
Plots lines and 95\% highest probability regions of the 2 dimensional models
specified by scales, slopes and correlations.
See help of 'line.models( )' for details about specifying the models.
}
\examples{
visualize.line.models(c(0.1,0.2), c(1,0), c(0.995,0))
}
